+++ src/main/java/next/career/domain/job/controller/JobController.java
@@ -6,6 +6,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import next.career.domain.job.controller.dto.GetJobDto;
 import next.career.domain.job.controller.dto.GetRoadMapDto;
+import next.career.domain.job.controller.dto.Work24;
 import next.career.domain.job.service.JobService;
 import next.career.domain.job.service.dto.JobDto;
 import next.career.domain.openai.dto.AiChatDto;
@@ -13,6 +14,7 @@ import next.career.domain.openai.dto.RecommendDto;
 import next.career.domain.user.entity.Member;
 import next.career.domain.user.entity.MemberDetail;
 import next.career.global.apiPayload.response.ApiResponse;
+import next.career.domain.job.service.HrdCourseService;
 import next.career.global.security.AuthDetails;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.data.domain.Page;
@@ -30,6 +32,7 @@ import java.util.Optional;
 public class JobController {
 
     private final JobService jobService;
+    private final HrdCourseService rawClient;
 
     // 전체 채용 조회
     @GetMapping("/all")
@@ -180,6 +183,15 @@ public class JobController {
         AiChatDto.HistoryResponse memberDetailResponse = AiChatDto.HistoryResponse.of(memberDetail);
 
         return ApiResponse.success(memberDetailResponse);
+    }
+
+    @GetMapping("/hrd-course")
+    public ApiResponse<Work24.CardCoursePage> raw(@RequestParam(defaultValue = "스프링") String keyword,
+                                                  @RequestParam(defaultValue = "1") int pageNo,
+                                                  @RequestParam(defaultValue = "10") int pageSize,
+                                                  @RequestParam(defaultValue = "20250101") String startYmd,
+                                                  @RequestParam(defaultValue = "20251231") String endYmd) {
+        return ApiResponse.success(rawClient.callRaw(keyword, pageNo, pageSize, startYmd, endYmd));
     }
 
 }
