+++ src/main/java/next/career/domain/job/service/HrdCourseService.java
@@ -1,17 +1,17 @@
-package next.career.global.client.restClient;
+package next.career.domain.job.service;
 
 import lombok.RequiredArgsConstructor;
+import next.career.domain.job.controller.dto.Work24;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClient;
 
 @Service
 @RequiredArgsConstructor
-public class Work24RawClient {
+public class HrdCourseService {
 
-    private final RestClient restClient;
+    private final RestClient work24Client;
 
     @Value("${work24.endpoints.card-course-list}")
     private String apiPath;
@@ -20,7 +20,7 @@ public class Work24RawClient {
     private String apiKey;
 
     public Work24.CardCoursePage callRaw(String keyword, int pageNo, int pageSize, String startYmd, String endYmd) {
-        return restClient.get()
+        return work24Client.get()
                 .uri(urlBuilder -> urlBuilder.path(apiPath)
                         .queryParam("authKey", apiKey)
                         .queryParam("returnType", "JSON")
